backend/requirements.txt
asgiref==3.9.1
boto3==1.40.12
botocore==1.40.12
cffi==1.17.1
cryptography==45.0.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
